LayoutTests/fast/url/data-uri-based-urls.html

<!DOCTYPE html>
<html>
<head>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpChildFramesAsText();
    }
</script>
</head>
<body>
<p>Test passes if i) the first iframe which includes a relative URL prints about:blank, ii) the second one which includes an absolute URL prints http://www.example.com, and iii) the third one which includes a fragment does not print about:blank</p>

<iframe src='data:text/html,<script>
document.write(`<a id="href" href="<p>FAILURE</p>"></a>`);
var src = document.getElementById("href").href;
document.write(src + "<br>");
</script>'></iframe>

<iframe src='data:text/html,<script>
document.write(`<a id="href" href="http://www.example.com"></a>`);
var src = document.getElementById("href").href;
document.write(src + "<br>");
</script>'></iframe>

<iframe src='data:text/html,<script>
document.write(`<a id="href" href="#fragment"></a>`);
var src = document.getElementById("href").href;
document.write(src + "<br>");
</script>'></iframe>
</body>
</html>
