LayoutTests/fast/url/data-uri-based-urls-expected.txt

Test passes if i) the first iframe which includes a relative URL prints about:blank, ii) the second one which includes an absolute URL prints http://www.example.com, and iii) the third one which includes a fragment does not print about:blank

   

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
about:blank


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
http://www.example.com/


--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
data:text/html,
